+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 0.4.11
 
-#### Apr 28, 2026
+#### Apr 29, 2026
 
 * [Feature] Added the ability to cancel unit improvements.
 * [Feature] Added a confirmation modal for cancelling unit improvements.
