+++ apps/web/app/(game)/(village-slug)/(village)/(...building-field-id)/components/components/smithy/smithy-unit-improvement.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { units } from '@pillage-first/game-assets/units';
 import { Bookmark } from 'app/(game)/(village-slug)/(village)/(...building-field-id)/components/components/bookmark';
@@ -12,22 +11,15 @@ import {
   Section,
   SectionContent,
 } from 'app/(game)/(village-slug)/components/building-layout';
-import { SmithyImprovementCancelDialog } from 'app/(game)/(village-slug)/components/smithy-improvement-cancel-dialog';
 import { SmithyImprovementTable } from 'app/(game)/(village-slug)/components/smithy-improvement-table';
 import { useTribe } from 'app/(game)/(village-slug)/hooks/use-tribe';
 import { useUnitResearch } from 'app/(game)/(village-slug)/hooks/use-unit-research';
 import { Text } from 'app/components/text';
-import { useDialog } from 'app/hooks/use-dialog';
 
 export const SmithyUnitImprovement = () => {
   const { t } = useTranslation();
   const tribe = useTribe();
   const { isUnitResearched } = useUnitResearch();
-  const {
-    isOpen: isCancelImprovementDialogOpen,
-    toggleModal: toggleCancelImprovementDialog,
-  } = useDialog();
-  const [eventIdToBeCancelled, setEventIdToBeCancelled] = useState<number>();
 
   const upgradableUnits = units.filter(({ category, tribe: unitTribe, id }) => {
     return (
@@ -36,11 +28,6 @@ export const SmithyUnitImprovement = () => {
       isUnitResearched(id)
     );
   });
-
-  const cancelUnitUpgrade = (upgradeEventId: number) => {
-    setEventIdToBeCancelled(upgradeEventId);
-    toggleCancelImprovementDialog();
-  };
 
   return (
     <Section>
@@ -54,7 +41,7 @@ export const SmithyUnitImprovement = () => {
         </Text>
       </SectionContent>
       <SectionContent>
-        <SmithyImprovementTable onImprovementCancel={cancelUnitUpgrade} />
+        <SmithyImprovementTable />
       </SectionContent>
       <SectionContent>
         {upgradableUnits.map(({ id }) => (
@@ -69,13 +56,6 @@ export const SmithyUnitImprovement = () => {
           </UnitCard>
         ))}
       </SectionContent>
-      {eventIdToBeCancelled && (
-        <SmithyImprovementCancelDialog
-          isOpen={isCancelImprovementDialogOpen}
-          onOpenChange={toggleCancelImprovementDialog}
-          eventId={eventIdToBeCancelled}
-        />
-      )}
     </Section>
   );
 };
